Display popup when a feature is clicked
The popup contains a table with information for the powerplants. A lot
of this stuff is taken from the "OpenLayers" popup example, so stuff
that's not working correctly (like the placing of the 'x' which closes
the popup) is related not having enough time to fix everything after
copy and pasting from the example.
The fact that to plants are clickable links (used as pseudo buttons)
prepares for future plotting functionality.