Add a marker on the center position of the map
The marker is a simple HTML <div> with a little bit of CSS that makes it
look like a dot. Since the CSS was a little bit longer I started to put
it into it's own dedicated file.
Using an overlay was the first thing that worked and since it's pretty
simple I'm going to stick with it for now.
Yes, this is the position of Magdeburg.