Optimize time series response
It's now a little bit better than quadratic complexity but there's still
room for improvement.
The limit was inserted because it doesn't make sense to display 47k time
series and it also took way too long.
As the names/ids for the powerplants are somewhat long in the 'dev'
schema, I'm using just a 'P' with incremental numbering as a suffix to
label the plots.