Inline `attrib` access and remove some `reduce`s
The `reduce` calls are no longer necessary and where only kept in the
code to make previous diffs more readable.

The `attrib` accesses should have been inlined from the get go. The only
reason I can think of that this didn't happen are line length issues,
which are gone now, since the code is a lot shorter anyway.