Fix dependency listing in the README
Be more specific concerning the type of the dependency. Currently they
are all **Python** packages.
Change 'geoalchemy' to 'geoalchemy2'. Of course we are trying to be up
to date and not use outdated packages.
Last but not leas: add 'psycopg2' to the list of dependencies. Since we're
going to use PostgreSQL/PostGIS anyways we're mostly going to rely on
this for database access.