Add an endpoint for the '/request_token' URI
While monitoring the communication between iD and the flask app, this is
the first URI that is requested, so it's the first one I'm adding. It is
covered pretty late in the oauthlib [tutorial][0] but that is because
all the machinery to make it work that is not in place yet is covered
beforehand.
Naturally requests leading to this endpoint fail for now, but I'll add
all the required machinery in subsequent commits.

[0]: http://flask-oauthlib.readthedocs.io/en/latest/oauth1.html#request-token-handler