General page style changes
The map no longer has a fixed size but should be as big as the viewport.
The plot is put below the map again, so the flow container can be
removed. As that means there no longer is a '.flow' element, the
corresponding style can be chucked out too.
The table with the database stats was just a test anyway and is gone
too. The title has been shortened and now hovers over the map. That
means it has to be the last thing in the HTML code as the map would
cover it otherwise.