Rename '_formhelpers.html' to 'macros.xml'
The original name was chosen by adhering to Flask's [WTForms
example][0]. As there will be other macros not connected to WTForms,
it's time to change the name. The new one, 'macros.xml', was chosen in
order to because currently it contains XML, of which XHTML is a dialect,
template code and until the macro functions can be divided in more
specific categories, 'macros' is a generic name, making the files
purpose clear.

[0]: http://flask.pocoo.org/docs/0.11/patterns/wtforms/#forms-in-templates