Add rudimentary config file parsing
In order to not have database access credentials in the code, they
should be in an external file. This file is now read on script startup
and the requirement for the existence of such a file and its format has
been duly noted in the README. Since this file should also never be
committed to the repository it has been put into '.gitignore' (which
thus had to be created).