Store created `Node`s in the database
This means we can finally get rid of the hackish in memory storing of
test data. Additionally, the ORM/database now takes care of assigning
ids. We just have to careful to always `add` and `commit` newly created
objects and everything should work out fine.
The indentation in `upload_changeset` is a little wonky, but that is
just done to keep as many lines untouched as possible, so that the diff
is a bit more readable.