Return correct stuff from getter/setter stubs
Now these functions work as expected. Note that some of the getters seem
to be intended to return `None` in case of missing objects to trigger
calling the corresponding setters. It's not documented when that is the
case and I don't know whether it's correct that way, but the way it is
now seems to work.