Implement the login machinery
Once again, this is pretty much by [the book][0] except that redirection
to `next` isn't implemented and we don't have to differentiate between
registration and login and. As this uses message flashing, the
corresponding code had to be added to the `login.html` template and
since the redirect on the `root` endpoint still isn't used properly (the
same redirect is hardwired into the `login` endpoint instead) a matching
TODO item has been added.

[0]: https://flask-login.readthedocs.io/en/latest/#login-example