Add `foreign_keys=[changeset_id]`
Looks like without this, SQLAlchemy can no longer figure out which
foreign key column to use to link changesets to elements. This might be
a result of the interaction between the introduction of the new
`tagged_id` foreign key and the `changeset_id` being declared using the
table name as a string due to the fact that the corresponding class is
declared later in the file.