Remove the 'md' and 'fl' overlays
This is one part of a general move away from overlays and towards vector
layers. The two overlays where just a hard coded proof of concept and
are no longer necessary. As the first release  will focus on
Schleswig-Holstein (SH) a point somewhere in the middle of SH is used as
the new initial center of the map's viewport.
The initial zoom level was also adjusted to be closer to the one in the
first release.