Add a second marker
This is a test/proof of concept that several markers based on one
template can be added. Doing this had a few consequences though.

The same marker can not be used twice so the template has to be cloned.
This means that the template snippet will not be relocated and thus has
to be put into an invisible "box", named 'outworld' here. The marker can
not be made itself invisible because than it would also not be displayed
on the map.

Since cloning creates elements sharing the same 'id' which is something
that shouldn't exist, the 'id' attributes had to be changed immediately
after cloning the template. This also means that styling can no longer
be done via id but has to be done via class.

And of course the variable names for the overlays had to be adjusted.