Put 'custom_widgets.js' under version control
This file currently contains some helper functions and a function to add
the scenario selection combobox to a parent container. Note that the
`context` parameter of `widgets.scenarios` currently isn't used, but has
been kept in the signature because it might be useful later to help with
not having to reload the whole page on a scenario change.