Add working database access
A little bit of a bigger commit but should be pretty self explanatory or
at least accessible together with the Flask and SQLAlchemy
documentation.
The database access should work and display the correct counts for the
objects in the two tables on the landing page (provided that the
connection is configured correctly).