Implement an endpoint for querying scenarios
The route is very simple and the output is a list of scenario names
contained in the database, serialized as a JSON string.
Getting the scenarios from the database could potentially be made more
efficient by refactoring the `tags` relation to use a dictionary instead
of a normal list. This would make other stuff easier to.