PEP 8
The style checker was complaining about the spaces after the angular
brackets, the empty line at the end of the file and some missing empty
lines in between. While I was at it, I also changed the styl (to many
empty lines) in the javascript file and fixed an unused imports
concerning the schemas.
But PEP 8 lets me use two spaces after the python imports, so I'm going
for that.