Update routes to changed iD behaviour
It seems that iD doesn't pick up the URL set in
`js/id/core/connection.js` when served from a relative URL. Or whatever
is the reason. It happens when it is served from the same server as the
openmod.sh app. That is actually a good thing, because it means it
doesn't pick up 'localhost' as the domain it is served from. The only
"downside" is that we have to change the routes accordingly, but that is
no big deal, as iD still seems to behave pretty straightforward.