Fix energiekollektiv/openmod.sh-id#2
Damn. Didn't test deleting a node which is part of a relation.
Also, this was sloppy coding. Copy and paste is bad enough. Getting bugs
because of not adjusting variable names afterwards just goes to show how
much of a better approach factoring this stuff out into reusable
functions should be.