First try at making `Node` an alias for `Element`
I'm trying to create an implementation where one can just use the names
of the classes used to communicate with the iD editor and they are
implicitly mapped to the underlying `Element` appropriately. The
committed file contains a first stab at this using ideas from [0] and
[1], but maybe using a [non-traditional mapping][2] would be more
appropriate?
The copied OSM classes are just here to have a quick reference on which
attributes are needed on their reimplementation.

[0]: http://docs.sqlalchemy.org/en/latest/orm/query.html#sqlalchemy.orm.query.Query.select_entity_from
[1]: http://docs.sqlalchemy.org/en/latest/orm/internals.html#sqlalchemy.orm.properties.RelationshipProperty.Comparator.any
[2]: http://docs.sqlalchemy.org/en/latest/orm/nonstandard_mappings.html