Comment out code destined for removal
This code will probably be removed later on. I commented it out, so that
it doesn't interfere with working code later on. I also didn't remove
it, because it will serve as a reminder on what has to be done to fully
implement the `map` api call.
The only part that was removed for good was the part which short
circuits if no nodes are found. Since the new design isn't node centric
anymore, there might still be contents which have to returned, even if
no nodes are found in the bounding box.
I also took the liberty of inserting some empty lines for better
readability.