Add and use a server side in memory session
This was necessary, since regular sessions have to be JSON serialisable
which `PointIds` was not. Now we are only storing a session id on the
client and are only storing everything else in a dictionary in memory.
If anything goes wrong with this implementation, most of the code was
taken from a [different implementation][0] of server side interfaces for
flask, so see that one for information on stuff which is missing.

[0]: https://github.com/fengsp/flask-session/blob/master/flask_session/sessions.py