Remove outdated comments
The information contained in the comments is on longer up to date.

The points of a geometry are accessed by first converting to a shapely
object and then using the respective shapely method to get the list of
points.

We don't use relations anymore, so there is no need to cater for
relations at all.

The points outside of the bounding box are automatically retrieved, as
ways are tested for intersection with the bounding box and all points
comprising the ways are returned.