Optimise bounding box intersection query
I don't really know whether this query is faster than the previous one,
but in principle it should be. It was also bugging me quite a bit that I
didn't find a good way to directly translate the simple:

  "check the scenario's children for intersection with the bounding box"

into sqlalchemy.

This query accomplishes that. It should be faster, since now we are not
searching every element for a parent with the current `scenario_id` but
are instead directly getting the children of the scenario with the known
id.

The necessary tricks where:

  a: use `session.query` to directly query the relationship, i.e.
     `Element.children`,
  b: use `aliased` to map the generated `subquery` to `Element` again
     and then
  c: tell sqlalchemy to use the `Element.geom` relationship to join with
     `Geom` since somehow the `aliased` call seems to mess up automatic
     detection of the foreign key relationship.