Fix stylesheet href in `base_layout.html` template
The fact that it's done relative means that it will only work in pages
whose URL is directly below the app's root URL, like e.g. all templates
in the `templates` directory. If one would have a template in a
subdirectory, stuff breaks.
Using an absolute URL here is a way safer approach.