Add the `iD.html` template
It displays the iD editor while respecting headers and footers and not
getting confused on where to place the iD editors widgets. Doing this
via an iframe is a bit of a hack though, as iframes are generally
frowned uppon. Maybe I can try a different way when I have more time.
All the other ones I tried didn't work, as the dynamically generated
floats in the iD editor where placed very weirdly.
I didn't override the `main.header`'s style to clear floats and have
dimensions of 100% with the other options though. As this was key to
getting things to work with the iframe, maybe it's also key to getting
things to work in a different way.
Also note: as I have to include iD specific scripts in the page and
override some style attributes of base template elements, I had to
change `base_layout.html` to have another block so that extending
templates can add stuff to the `<head>` element.