Don't destroy session under mapped object
The mapped object returned from `get_elements` should belong to a
session. If the session is created in `get_elements` it is destroyed
when the function returns but the object pulled out of the session is
still alive. This can lead to all sorts of problems, so the session is
now created in the calling function and passed down appropriately.