Make `api.py` session handling safer
Making session creation a proper `contextmanager` that cleans up and
closes the session after the code using the session is run, takes care
of all database lockups. The lockups where created by stray session not
destroyed by the garbage collector and thus still open.
The commit looks huge, but this is only because all the code that got
put under the `with db_session() as session:` had to be indented by one
more level, leading to a lot of changes that are actually just more
leading whitespace and some code reformatting due to now overly long
lines.