Fully put `get_hub_results` into a session context
I didn't put the whole function into the session context, as I didn't
think it was necessary. Turns out it actually is, as objects queried
using the session are used throughout the whole function.
So we either gotta keep the whole session alive throughout the function,
or load the relationships and columns eagerly.