Differentiate between the jobs page and widget
The widget, which is just a `<div>` containing hte list of jobs, can be
`{% include %}`ed wherever it fits, while the page can be rendered
standalone. The page actually just extends the `base_layout.html` by
including the widget.
In order to do this, I only had to make a new route and template for the
page and rewire the `Jobs` navigation link.