Remove the jobs widget from `edit_scenario.html`
Information about the jobs now has it's own page, so there's no need to
display it when editing the scenario anymore. That also means that the
`jobs` argument for the `edit_scenario.html` template is no longer
necessary.
I left the success callback for the `runSimulation` XHR empty
for refernce purposes.