Introduce and `Update` button
Now we can update the job list without reloading the whole page.
This is important because if you don't update the view before removing
dead jobs, jobs which are not yet marked as `Done.`, but are already
done on the server might disappear surprisingly.
This is actually just a quick hack and can be removed, once I figure out
an elegant way of having life updates via long polling or a long lived
request with a streaming response.