Change how jobs are displayed in the jobs widget
Now we are displaying a "Scenario Name (Owning User)" string as the link
text which is much more informative than the opaque job key.
Also, iterating over `job.items()` actually makes the code shorter AND
more readable as we now don't have warts like `jobs[job]` anymore and
`job` is actually a job object and not a string. Hooray!