Change job stop signal to SIGUSR2
This was an accident waiting to happen. Sorry for that.
SIGINT is so ubiquitious that Apache2 uses it and so it should be
avoided. SIGUSR2 should be safe. Signal handling has to be enabled using
`WSGIRestrictSignal Off` in the Apache2 configuration though.
In the long run, I should find a way of stopping running jobs without
sending a signal to them. Maybe using a pipe and sending data to that.
Although setting up asynchronous, possibly evented I/O on both, server
and worker process side is something I've been banging my head against
for a while, it seems like there is no way around it. *sigh*