Remove erroneous `gettext`
This time the correct function, i.e. `gettext` instead of `getText` was
used, but this is the wrong spot to do localization.
First, we're not in flask land. This file is run in a process that
doesn't know anything about flask or the fact that it's part of a web
application. This is basically just stuff which is run as a regular
python program, so we shouldn't depend on `flask_babel` here.
Second, you should rather do localization right before displaying the
description string and not here.
That's why removing the call is the right solution in this case.