Display 'job started' dialog at the correct time
This was wonky. The dialog stating that the job now runs in the
background was displayed first, and the upload starting the job was
actually started by the callback that is evaluated after the dialog is
hidden again. That's the wrong way around.
Now the job upload is started immediately and the dialog is displayed
after the upload is finished.